Module & Scene switch refactoring
The way modules are defined was completely updated with a static layer handling base data and GameObjects handling interface with the scene.
That way scene switch when diving could also be refactored to avoid a physical scene swith in Unity. Now, everything happens in a unique scene and diving is "just" a matter of moving module data from the list of visible modules to the list of stored module (the bank).
In the process user interaction was also updated to better handle module importation in the scene. Modules dedicated to connecting to a remote http server were added.
The possibility to dive out of a module has not yet been reintroduced.